Package Structure
Fixes #55
