+++ playground/chatbot.py
@@ -1,0 +1,102 @@
+from langchain.chains.combine_documents import create_stuff_documents_chain
+from langchain_core.chat_history import BaseChatMessageHistory
+from langchain.chains.history_aware_retriever import create_history_aware_retriever
+from langchain_core.runnables.history import RunnableWithMessageHistory
+from langchain.chains.retrieval import create_retrieval_chain
+from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
+from langchain_unify.chat_models import ChatUnify
+from playground import st
+from playground.document_processing import process_inputs, format_docs, output_chunks
+from langchain_community.chat_message_histories import ChatMessageHistory
+
+
+def ask_unify():
+    if "vector_store" not in st.session_state:
+        process_inputs()
+
+    retriever = st.session_state.vector_store.as_retriever()
+
+    if "model_temperature" not in st.session_state:
+        st.session_state.model_temperature = 0.3
+
+    model = ChatUnify(
+        model=st.session_state.endpoint,
+        unify_api_key=st.session_state.unify_api_key,
+        temperature=st.session_state.model_temperature
+    )
+
+    contextualize_q_system_prompt = """Given a chat history and the latest user question \
+    which might reference context in the chat history, formulate a standalone question \
+    which can be understood without the chat history. Do NOT answer the question, \
+    just reformulate it if needed and otherwise return it as is."""
+    contextualize_q_prompt = ChatPromptTemplate.from_messages(
+        [
+            ("system", contextualize_q_system_prompt),
+            MessagesPlaceholder("chat_history"),
+            ("human", "{input}"),
+        ]
+    )
+
+    history_aware_retriever = create_history_aware_retriever(
+        model, retriever | format_docs, contextualize_q_prompt
+    )
+
+    qa_system_prompt = """You are an assistant for question-answering tasks. \
+    Use the following pieces of retrieved context to answer the question. \
+    If you don't know the answer, just say that you don't know. \
+    Use three sentences maximum and keep the answer concise.\
+
+    {context}"""
+    qa_prompt = ChatPromptTemplate.from_messages(
+        [
+            ("system", qa_system_prompt),
+            MessagesPlaceholder("chat_history"),
+            ("human", "{input}"),
+        ]
+    )
+    question_answer_chain = create_stuff_documents_chain(model, qa_prompt)
+
+    rag_chain = create_retrieval_chain(history_aware_retriever, question_answer_chain)
+
+    if "store" not in st.session_state:
+        st.session_state.store = {}
+
+    def get_session_history(session_id: str) -> BaseChatMessageHistory:
+        if session_id not in st.session_state.store:
+            st.session_state.store[session_id] = ChatMessageHistory()
+        return st.session_state.store[session_id]
+
+    conversational_rag_chain = RunnableWithMessageHistory(
+        rag_chain,
+        get_session_history,
+        input_messages_key="input",
+        history_messages_key="chat_history",
+        output_messages_key="answer"
+    )
+
+    return conversational_rag_chain
+
+
+def chat_bot():
+    if "messages" not in st.session_state:
+        st.session_state.messages = []
+    #
+    for message in st.session_state.messages:
+        st.chat_message('human').write(message[0])
+        st.chat_message('assistant').write(message[1])
+    #
+    if query := st.chat_input("Ask your document anything...", key="query"):
+
+        if "processed_input" not in st.session_state:
+            st.warning("Please input your details in the sidebar first")
+            return
+
+        st.chat_message("human").write(query)
+
+        conversational_rag_chain = ask_unify()
+
+        response = st.chat_message("assistant").write_stream(
+            output_chunks(conversational_rag_chain, query)
+        )
+
+        st.session_state.messages.append((query, response))
