+++ requirements.txt
@@ -1,4 +1,4 @@
-streamlit
+streamlit>=1.34.0
 PyPDF2
 sentence_transformers
 faiss-cpu
