+++ run.py
@@ -1,0 +1,5 @@
+from playground import landing_page
+
+
+if __name__ == "__main__":
+    landing_page()
